Create STL
 In our second class of the boot camp, we delved into the fundamental concepts of the Standard Template Library (STL), a powerful set of C++ template classes to provide common programming data structures and functions such as lists, stacks, queues, and algorithms.

Throughout the session, we explored the versatility and efficiency of STL in solving various competitive programming problems. From understanding the syntax to implementing complex algorithms, we've embarked on a journey to harness the full potential of STL.

This folder serves as a repository for all the editorials and code snippets covered during our exploration of STL. Here, you'll find detailed explanations, insightful tips, and well-commented code samples that will help you understand STL concepts.